TAL-239 | Android | Sign In
Worked on the second screen for the Sign In. When opening the app, the Splash will start the activity in which you enter an organization's id, after tapping next the activity in which you enter email and password will start.
TODO: functionality, shadows on some EditTexts, clickable links in some TextViews.